changed centroid to lat lon

+++ db/migrate/20131028225506_remove_centroid_from_polygon.rb
@@ -1,7 +1,7 @@
 class RemoveCentroidFromPolygon < ActiveRecord::Migration
   def change
   	remove_column :polygons, :centroid
-  	add_column :polygons, :centroid_lat, :double
-  	add_column :polygons, :centroid_lon, :double
+  	add_column :polygons, :centroid_lat, :float
+  	add_column :polygons, :centroid_lon, :float
   end
 end
